pyproject.toml
[build-system]
requires = ["flit_core >=3.11,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "message_sender_telegram_bot"
authors = [{ name = "roli2py", email = "romanmashevskyi@proton.me" }]
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = "~=3.14"
dynamic = ["version", "description"]
dependencies = [
    "python-telegram-bot ~= 22.5",
    "sqlalchemy ~= 2.0.45",
    "mysqlclient ~= 2.2.7",
    "alembic ~= 1.17.2",
]

[project.optional-dependencies]
test = ["pytest ~= 9.0.2", "pytest-cov ~= 7.0.0"]
dev = [
    "pytest ~= 9.0.2",
    "pytest-cov ~= 7.0.0",
    "isort ~= 7.0.0",
    "black ~= 25.12.0",
]

[project.urls]
Home = "https://github.com/roli2py/message-sender-telegram-bot"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
pythonpath = ["./src/message_sender_telegram_bot"]

[[tool.basedpyright.executionEnvironments]]
root = "./src/message_sender_telegram_bot"

[[tool.basedpyright.executionEnvironments]]
root = "./tests"
extraPaths = ["./src/message_sender_telegram_bot"]

[tool.isort]
profile = "black"
skip_gitignore = true
src_paths = ["./src/message_sender_telegram_bot"]

[tool.black]
line-length = 79
